Complete logic and style with all FCC tests passed
Worked out a lot of the niggles involving using decimal points and the
bizarre requirement of the conditional use of the subtract symbol as a
negative key. Would clearly be much easier to implement a +/- toggle.
Styling was fiddly using Bootstrap but eventually realised to align
inline elements I needed to treat it as text. Still not ideal from a
best practice point of view. A lot of clunky code, but good enough
for now.